rg35xxsp: single-FAT32 boot, BlitContext, clippy clean, wifi stubs, u-boot flash script
- Switch boot to single-FAT32 with initramfs loop-mounting system.erofs
- boot.cmd: fatload + booti with uInitrd; uboot.config: enable FAT cmds
- rg35xxsp.mk: deploy via rsync (no disk image); add flash-bootloader.command
- initramfs: mounts mmcblk0p1 as FAT32, loop-mounts system.erofs, switch_root
- openrc-services.conf: drop iwd/wifi-net-sp; base.world: wpa_supplicant
- Add BlitContext struct bundling per-blit params (mimics GFX_Renderer)
- Collapse 8-10 arg row fns into 5-arg; remove too_many_arguments allows
- Add RG35xxSP wifi support: modprobe, wpa_supplicant, udhcpc, ip parsing
- Remove 12+ stale #[allow(dead_code)] attrs from play crate
- Fix all play warnings: unused imports, dead code, collapsible let, clamp
- Update local.mk: SDCARD_PATH for /Volumes/ALLIUM

+++ crates/play/src/video/effects.rs
@@ -11,25 +11,23 @@ pub fn weight3_1_rgb565(a: u16, b: u16) -> u16 {
     let r = ((3 * ((a >> 11) & 0x1F) + ((b >> 11) & 0x1F)) >> 2) << 11;
     let g = ((3 * ((a >> 5) & 0x3F) + ((b >> 5) & 0x3F)) >> 2) << 5;
     let b_out = (3 * (a & 0x1F) + (b & 0x1F)) >> 2;
-    (r | g | b_out) as u16
+    r | g | b_out
 }
 
 /// Blend two RGB565 pixels with weight 2:3 (a gets 40%, b gets 60%).
-#[allow(dead_code)]
 pub fn weight2_3_rgb565(a: u16, b: u16) -> u16 {
     let r = ((2 * ((a >> 11) & 0x1F) + 3 * ((b >> 11) & 0x1F)) / 5) << 11;
     let g = ((2 * ((a >> 5) & 0x3F) + 3 * ((b >> 5) & 0x3F)) / 5) << 5;
     let b_out = (2 * (a & 0x1F) + 3 * (b & 0x1F)) / 5;
-    (r | g | b_out) as u16
+    r | g | b_out
 }
 
 /// Blend two RGB565 pixels with weight 3:2 (a gets 60%, b gets 40%).
-#[allow(dead_code)]
 pub fn weight3_2_rgb565(a: u16, b: u16) -> u16 {
     let r = ((3 * ((a >> 11) & 0x1F) + 2 * ((b >> 11) & 0x1F)) / 5) << 11;
     let g = ((3 * ((a >> 5) & 0x3F) + 2 * ((b >> 5) & 0x3F)) / 5) << 5;
     let b_out = (3 * (a & 0x1F) + 2 * (b & 0x1F)) / 5;
-    (r | g | b_out) as u16
+    r | g | b_out
 }
 
 /// Apply a procedural effect to an RGB565 pixel at a given output position.
